services/auth/requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.7
bcrypt==4.1.3
PyJWT==2.8.0
boto3==1.35.23
marshmallow==3.22.0
python-dotenv==1.0.1
pyotp==2.9.0
